Fixing to use with postman

+++ API/deleteContact.php
@@ -2,10 +2,18 @@
 include "../conn.php";
 session_start();
 
-if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-  $contact_id = $_GET["contact_id"];
+function isPostmanRequest()
+{
+    return isset($_SERVER['HTTP_USER_AGENT']) && strpos($_SERVER['HTTP_USER_AGENT'], 'Postman') !== false;
+}
 
-  $sql = "DELETE FROM `contacts` WHERE `id` = '$contact_id'";
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+  if (isPostmanRequest()) {
+    echo "POST request received\n";
+  }
+  $user_id = $_SESSION['id'];
+  $contact_id = $_GET["contact_id"];
+  $sql = "DELETE FROM `contacts` WHERE `id` = '$contact_id' AND `user_id` = '$user_id'";
   $result = mysqli_query($conn, $sql);
   
   if ($result) {
@@ -14,5 +22,8 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     echo "Failed: " . mysqli_error($conn);
     header("Location: ../dashboard.php");
   }
+}
+else {
+  echo "<script> window.location='../dashboard.php';</script>";
 }
 ?>
